requirements.txt
streamlit
openai
openai-whisper
torch
ffmpeg-python
python-dotenv
